Flatten the design of vk-sys
