--- vk-sys/src/enums.rs
@@ -1,944 +1,0 @@
-// Copyright (c) 2016 The vulkano developers
-// Licensed under the Apache License, Version 2.0
-// <LICENSE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT
-// license <LICENSE-MIT or http://opensource.org/licenses/MIT>,
-// at your option. All files in the project carrying such
-// notice may not be copied, modified, or distributed except
-// according to those terms.
-
-
-pub type Flags = u32;
-pub type Bool32 = u32;
-pub type DeviceSize = u64;
-pub type SampleMask = u32;
-
-pub type Instance = usize;
-pub type PhysicalDevice = usize;
-pub type Device = usize;
-pub type Queue = usize;
-pub type CommandBuffer = usize;
-
-pub type Semaphore = u64;
-pub type Fence = u64;
-pub type DeviceMemory = u64;
-pub type Buffer = u64;
-pub type Image = u64;
-pub type Event = u64;
-pub type QueryPool = u64;
-pub type BufferView = u64;
-pub type ImageView = u64;
-pub type ShaderModule = u64;
-pub type PipelineCache = u64;
-pub type PipelineLayout = u64;
-pub type RenderPass = u64;
-pub type Pipeline = u64;
-pub type DescriptorSetLayout = u64;
-pub type Sampler = u64;
-pub type DescriptorPool = u64;
-pub type DescriptorSet = u64;
-pub type Framebuffer = u64;
-pub type CommandPool = u64;
-pub type SurfaceKHR = u64;
-pub type SwapchainKHR = u64;
-pub type DisplayKHR = u64;
-pub type DisplayModeKHR = u64;
-pub type DebugReportCallbackEXT = u64;
-
-pub const LOD_CLAMP_NONE: f32 = 1000.0;
-pub const REMAINING_MIP_LEVELS: u32 = 0xffffffff;
-pub const REMAINING_ARRAY_LAYERS: u32 = 0xffffffff;
-pub const WHOLE_SIZE: u64 = 0xffffffffffffffff;
-pub const ATTACHMENT_UNUSED: u32 = 0xffffffff;
-pub const TRUE: u32 = 1;
-pub const FALSE: u32 = 0;
-pub const QUEUE_FAMILY_IGNORED: u32 = 0xffffffff;
-pub const SUBPASS_EXTERNAL: u32 = 0xffffffff;
-pub const MAX_PHYSICAL_DEVICE_NAME_SIZE: u32 = 256;
-pub const UUID_SIZE: u32 = 16;
-pub const MAX_MEMORY_TYPES: u32 = 32;
-pub const MAX_MEMORY_HEAPS: u32 = 16;
-pub const MAX_EXTENSION_NAME_SIZE: u32 = 256;
-pub const MAX_DESCRIPTION_SIZE: u32 = 256;
-
-pub type PipelineCacheHeaderVersion = u32;
-pub const PIPELINE_CACHE_HEADER_VERSION_ONE: u32 = 1;
-
-pub type Result = u32;
-pub const SUCCESS: u32 = 0;
-pub const NOT_READY: u32 = 1;
-pub const TIMEOUT: u32 = 2;
-pub const EVENT_SET: u32 = 3;
-pub const EVENT_RESET: u32 = 4;
-pub const INCOMPLETE: u32 = 5;
-pub const ERROR_OUT_OF_HOST_MEMORY: u32 = -1i32 as u32;
-pub const ERROR_OUT_OF_DEVICE_MEMORY: u32 = -2i32 as u32;
-pub const ERROR_INITIALIZATION_FAILED: u32 = -3i32 as u32;
-pub const ERROR_DEVICE_LOST: u32 = -4i32 as u32;
-pub const ERROR_MEMORY_MAP_FAILED: u32 = -5i32 as u32;
-pub const ERROR_LAYER_NOT_PRESENT: u32 = -6i32 as u32;
-pub const ERROR_EXTENSION_NOT_PRESENT: u32 = -7i32 as u32;
-pub const ERROR_FEATURE_NOT_PRESENT: u32 = -8i32 as u32;
-pub const ERROR_INCOMPATIBLE_DRIVER: u32 = -9i32 as u32;
-pub const ERROR_TOO_MANY_OBJECTS: u32 = -10i32 as u32;
-pub const ERROR_FORMAT_NOT_SUPPORTED: u32 = -11i32 as u32;
-pub const ERROR_SURFACE_LOST_KHR: u32 = -1000000000i32 as u32;
-pub const ERROR_NATIVE_WINDOW_IN_USE_KHR: u32 = -1000000001i32 as u32;
-pub const SUBOPTIMAL_KHR: u32 = 1000001003;
-pub const ERROR_OUT_OF_DATE_KHR: u32 = -1000001004i32 as u32;
-pub const ERROR_INCOMPATIBLE_DISPLAY_KHR: u32 = -1000003001i32 as u32;
-pub const ERROR_VALIDATION_FAILED_EXT: u32 = -1000011001i32 as u32;
-
-pub type StructureType = u32;
-pub const STRUCTURE_TYPE_APPLICATION_INFO: u32 = 0;
-pub const STRUCTURE_TYPE_INSTANCE_CREATE_INFO: u32 = 1;
-pub const STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO: u32 = 2;
-pub const STRUCTURE_TYPE_DEVICE_CREATE_INFO: u32 = 3;
-pub const STRUCTURE_TYPE_SUBMIT_INFO: u32 = 4;
-pub const STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO: u32 = 5;
-pub const STRUCTURE_TYPE_MAPPED_MEMORY_RANGE: u32 = 6;
-pub const STRUCTURE_TYPE_BIND_SPARSE_INFO: u32 = 7;
-pub const STRUCTURE_TYPE_FENCE_CREATE_INFO: u32 = 8;
-pub const STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO: u32 = 9;
-pub const STRUCTURE_TYPE_EVENT_CREATE_INFO: u32 = 10;
-pub const STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO: u32 = 11;
-pub const STRUCTURE_TYPE_BUFFER_CREATE_INFO: u32 = 12;
-pub const STRUCTURE_TYPE_BUFFER_VIEW_CREATE_INFO: u32 = 13;
-pub const STRUCTURE_TYPE_IMAGE_CREATE_INFO: u32 = 14;
-pub const STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO: u32 = 15;
-pub const STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO: u32 = 16;
-pub const STRUCTURE_TYPE_PIPELINE_CACHE_CREATE_INFO: u32 = 17;
-pub const STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO: u32 = 18;
-pub const STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO: u32 = 19;
-pub const STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO: u32 = 20;
-pub const STRUCTURE_TYPE_PIPELINE_TESSELLATION_STATE_CREATE_INFO: u32 = 21;
-pub const STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO: u32 = 22;
-pub const STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO: u32 = 23;
-pub const STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO: u32 = 24;
-pub const STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO: u32 = 25;
-pub const STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO: u32 = 26;
-pub const STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO: u32 = 27;
-pub const STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO: u32 = 28;
-pub const STRUCTURE_TYPE_COMPUTE_PIPELINE_CREATE_INFO: u32 = 29;
-pub const STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO: u32 = 30;
-pub const STRUCTURE_TYPE_SAMPLER_CREATE_INFO: u32 = 31;
-pub const STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO: u32 = 32;
-pub const STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO: u32 = 33;
-pub const STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO: u32 = 34;
-pub const STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET: u32 = 35;
-pub const STRUCTURE_TYPE_COPY_DESCRIPTOR_SET: u32 = 36;
-pub const STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO: u32 = 37;
-pub const STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO: u32 = 38;
-pub const STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO: u32 = 39;
-pub const STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO: u32 = 40;
-pub const STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_INFO: u32 = 41;
-pub const STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO: u32 = 42;
-pub const STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO: u32 = 43;
-pub const STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER: u32 = 44;
-pub const STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER: u32 = 45;
-pub const STRUCTURE_TYPE_MEMORY_BARRIER: u32 = 46;
-pub const STRUCTURE_TYPE_LOADER_INSTANCE_CREATE_INFO: u32 = 47;
-pub const STRUCTURE_TYPE_LOADER_DEVICE_CREATE_INFO: u32 = 48;
-pub const STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR: u32 = 1000001000;
-pub const STRUCTURE_TYPE_PRESENT_INFO_KHR: u32 = 1000001001;
-pub const STRUCTURE_TYPE_DISPLAY_MODE_CREATE_INFO_KHR: u32 = 1000002000;
-pub const STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR: u32 = 1000002001;
-pub const STRUCTURE_TYPE_DISPLAY_PRESENT_INFO_KHR: u32 = 1000003000;
-pub const STRUCTURE_TYPE_XLIB_SURFACE_CREATE_INFO_KHR: u32 = 1000004000;
-pub const STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR: u32 = 1000005000;
-pub const STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR: u32 = 1000006000;
-pub const STRUCTURE_TYPE_MIR_SURFACE_CREATE_INFO_KHR: u32 = 1000007000;
-pub const STRUCTURE_TYPE_ANDROID_SURFACE_CREATE_INFO_KHR: u32 = 1000008000;
-pub const STRUCTURE_TYPE_WIN32_SURFACE_CREATE_INFO_KHR: u32 = 1000009000;
-pub const STRUCTURE_TYPE_DEBUG_REPORT_CREATE_INFO_EXT: u32 = 1000011000;
-
-pub type SystemAllocationScope = u32;
-pub const SYSTEM_ALLOCATION_SCOPE_COMMAND: u32 = 0;
-pub const SYSTEM_ALLOCATION_SCOPE_OBJECT: u32 = 1;
-pub const SYSTEM_ALLOCATION_SCOPE_CACHE: u32 = 2;
-pub const SYSTEM_ALLOCATION_SCOPE_DEVICE: u32 = 3;
-pub const SYSTEM_ALLOCATION_SCOPE_INSTANCE: u32 = 4;
-
-pub type InternalAllocationType = u32;
-pub const INTERNAL_ALLOCATION_TYPE_EXECUTABLE: u32 = 0;
-
-pub type Format = u32;
-pub const FORMAT_UNDEFINED: u32 = 0;
-pub const FORMAT_R4G4_UNORM_PACK8: u32 = 1;
-pub const FORMAT_R4G4B4A4_UNORM_PACK16: u32 = 2;
-pub const FORMAT_B4G4R4A4_UNORM_PACK16: u32 = 3;
-pub const FORMAT_R5G6B5_UNORM_PACK16: u32 = 4;
-pub const FORMAT_B5G6R5_UNORM_PACK16: u32 = 5;
-pub const FORMAT_R5G5B5A1_UNORM_PACK16: u32 = 6;
-pub const FORMAT_B5G5R5A1_UNORM_PACK16: u32 = 7;
-pub const FORMAT_A1R5G5B5_UNORM_PACK16: u32 = 8;
-pub const FORMAT_R8_UNORM: u32 = 9;
-pub const FORMAT_R8_SNORM: u32 = 10;
-pub const FORMAT_R8_USCALED: u32 = 11;
-pub const FORMAT_R8_SSCALED: u32 = 12;
-pub const FORMAT_R8_UINT: u32 = 13;
-pub const FORMAT_R8_SINT: u32 = 14;
-pub const FORMAT_R8_SRGB: u32 = 15;
-pub const FORMAT_R8G8_UNORM: u32 = 16;
-pub const FORMAT_R8G8_SNORM: u32 = 17;
-pub const FORMAT_R8G8_USCALED: u32 = 18;
-pub const FORMAT_R8G8_SSCALED: u32 = 19;
-pub const FORMAT_R8G8_UINT: u32 = 20;
-pub const FORMAT_R8G8_SINT: u32 = 21;
-pub const FORMAT_R8G8_SRGB: u32 = 22;
-pub const FORMAT_R8G8B8_UNORM: u32 = 23;
-pub const FORMAT_R8G8B8_SNORM: u32 = 24;
-pub const FORMAT_R8G8B8_USCALED: u32 = 25;
-pub const FORMAT_R8G8B8_SSCALED: u32 = 26;
-pub const FORMAT_R8G8B8_UINT: u32 = 27;
-pub const FORMAT_R8G8B8_SINT: u32 = 28;
-pub const FORMAT_R8G8B8_SRGB: u32 = 29;
-pub const FORMAT_B8G8R8_UNORM: u32 = 30;
-pub const FORMAT_B8G8R8_SNORM: u32 = 31;
-pub const FORMAT_B8G8R8_USCALED: u32 = 32;
-pub const FORMAT_B8G8R8_SSCALED: u32 = 33;
-pub const FORMAT_B8G8R8_UINT: u32 = 34;
-pub const FORMAT_B8G8R8_SINT: u32 = 35;
-pub const FORMAT_B8G8R8_SRGB: u32 = 36;
-pub const FORMAT_R8G8B8A8_UNORM: u32 = 37;
-pub const FORMAT_R8G8B8A8_SNORM: u32 = 38;
-pub const FORMAT_R8G8B8A8_USCALED: u32 = 39;
-pub const FORMAT_R8G8B8A8_SSCALED: u32 = 40;
-pub const FORMAT_R8G8B8A8_UINT: u32 = 41;
-pub const FORMAT_R8G8B8A8_SINT: u32 = 42;
-pub const FORMAT_R8G8B8A8_SRGB: u32 = 43;
-pub const FORMAT_B8G8R8A8_UNORM: u32 = 44;
-pub const FORMAT_B8G8R8A8_SNORM: u32 = 45;
-pub const FORMAT_B8G8R8A8_USCALED: u32 = 46;
-pub const FORMAT_B8G8R8A8_SSCALED: u32 = 47;
-pub const FORMAT_B8G8R8A8_UINT: u32 = 48;
-pub const FORMAT_B8G8R8A8_SINT: u32 = 49;
-pub const FORMAT_B8G8R8A8_SRGB: u32 = 50;
-pub const FORMAT_A8B8G8R8_UNORM_PACK32: u32 = 51;
-pub const FORMAT_A8B8G8R8_SNORM_PACK32: u32 = 52;
-pub const FORMAT_A8B8G8R8_USCALED_PACK32: u32 = 53;
-pub const FORMAT_A8B8G8R8_SSCALED_PACK32: u32 = 54;
-pub const FORMAT_A8B8G8R8_UINT_PACK32: u32 = 55;
-pub const FORMAT_A8B8G8R8_SINT_PACK32: u32 = 56;
-pub const FORMAT_A8B8G8R8_SRGB_PACK32: u32 = 57;
-pub const FORMAT_A2R10G10B10_UNORM_PACK32: u32 = 58;
-pub const FORMAT_A2R10G10B10_SNORM_PACK32: u32 = 59;
-pub const FORMAT_A2R10G10B10_USCALED_PACK32: u32 = 60;
-pub const FORMAT_A2R10G10B10_SSCALED_PACK32: u32 = 61;
-pub const FORMAT_A2R10G10B10_UINT_PACK32: u32 = 62;
-pub const FORMAT_A2R10G10B10_SINT_PACK32: u32 = 63;
-pub const FORMAT_A2B10G10R10_UNORM_PACK32: u32 = 64;
-pub const FORMAT_A2B10G10R10_SNORM_PACK32: u32 = 65;
-pub const FORMAT_A2B10G10R10_USCALED_PACK32: u32 = 66;
-pub const FORMAT_A2B10G10R10_SSCALED_PACK32: u32 = 67;
-pub const FORMAT_A2B10G10R10_UINT_PACK32: u32 = 68;
-pub const FORMAT_A2B10G10R10_SINT_PACK32: u32 = 69;
-pub const FORMAT_R16_UNORM: u32 = 70;
-pub const FORMAT_R16_SNORM: u32 = 71;
-pub const FORMAT_R16_USCALED: u32 = 72;
-pub const FORMAT_R16_SSCALED: u32 = 73;
-pub const FORMAT_R16_UINT: u32 = 74;
-pub const FORMAT_R16_SINT: u32 = 75;
-pub const FORMAT_R16_SFLOAT: u32 = 76;
-pub const FORMAT_R16G16_UNORM: u32 = 77;
-pub const FORMAT_R16G16_SNORM: u32 = 78;
-pub const FORMAT_R16G16_USCALED: u32 = 79;
-pub const FORMAT_R16G16_SSCALED: u32 = 80;
-pub const FORMAT_R16G16_UINT: u32 = 81;
-pub const FORMAT_R16G16_SINT: u32 = 82;
-pub const FORMAT_R16G16_SFLOAT: u32 = 83;
-pub const FORMAT_R16G16B16_UNORM: u32 = 84;
-pub const FORMAT_R16G16B16_SNORM: u32 = 85;
-pub const FORMAT_R16G16B16_USCALED: u32 = 86;
-pub const FORMAT_R16G16B16_SSCALED: u32 = 87;
-pub const FORMAT_R16G16B16_UINT: u32 = 88;
-pub const FORMAT_R16G16B16_SINT: u32 = 89;
-pub const FORMAT_R16G16B16_SFLOAT: u32 = 90;
-pub const FORMAT_R16G16B16A16_UNORM: u32 = 91;
-pub const FORMAT_R16G16B16A16_SNORM: u32 = 92;
-pub const FORMAT_R16G16B16A16_USCALED: u32 = 93;
-pub const FORMAT_R16G16B16A16_SSCALED: u32 = 94;
-pub const FORMAT_R16G16B16A16_UINT: u32 = 95;
-pub const FORMAT_R16G16B16A16_SINT: u32 = 96;
-pub const FORMAT_R16G16B16A16_SFLOAT: u32 = 97;
-pub const FORMAT_R32_UINT: u32 = 98;
-pub const FORMAT_R32_SINT: u32 = 99;
-pub const FORMAT_R32_SFLOAT: u32 = 100;
-pub const FORMAT_R32G32_UINT: u32 = 101;
-pub const FORMAT_R32G32_SINT: u32 = 102;
-pub const FORMAT_R32G32_SFLOAT: u32 = 103;
-pub const FORMAT_R32G32B32_UINT: u32 = 104;
-pub const FORMAT_R32G32B32_SINT: u32 = 105;
-pub const FORMAT_R32G32B32_SFLOAT: u32 = 106;
-pub const FORMAT_R32G32B32A32_UINT: u32 = 107;
-pub const FORMAT_R32G32B32A32_SINT: u32 = 108;
-pub const FORMAT_R32G32B32A32_SFLOAT: u32 = 109;
-pub const FORMAT_R64_UINT: u32 = 110;
-pub const FORMAT_R64_SINT: u32 = 111;
-pub const FORMAT_R64_SFLOAT: u32 = 112;
-pub const FORMAT_R64G64_UINT: u32 = 113;
-pub const FORMAT_R64G64_SINT: u32 = 114;
-pub const FORMAT_R64G64_SFLOAT: u32 = 115;
-pub const FORMAT_R64G64B64_UINT: u32 = 116;
-pub const FORMAT_R64G64B64_SINT: u32 = 117;
-pub const FORMAT_R64G64B64_SFLOAT: u32 = 118;
-pub const FORMAT_R64G64B64A64_UINT: u32 = 119;
-pub const FORMAT_R64G64B64A64_SINT: u32 = 120;
-pub const FORMAT_R64G64B64A64_SFLOAT: u32 = 121;
-pub const FORMAT_B10G11R11_UFLOAT_PACK32: u32 = 122;
-pub const FORMAT_E5B9G9R9_UFLOAT_PACK32: u32 = 123;
-pub const FORMAT_D16_UNORM: u32 = 124;
-pub const FORMAT_X8_D24_UNORM_PACK32: u32 = 125;
-pub const FORMAT_D32_SFLOAT: u32 = 126;
-pub const FORMAT_S8_UINT: u32 = 127;
-pub const FORMAT_D16_UNORM_S8_UINT: u32 = 128;
-pub const FORMAT_D24_UNORM_S8_UINT: u32 = 129;
-pub const FORMAT_D32_SFLOAT_S8_UINT: u32 = 130;
-pub const FORMAT_BC1_RGB_UNORM_BLOCK: u32 = 131;
-pub const FORMAT_BC1_RGB_SRGB_BLOCK: u32 = 132;
-pub const FORMAT_BC1_RGBA_UNORM_BLOCK: u32 = 133;
-pub const FORMAT_BC1_RGBA_SRGB_BLOCK: u32 = 134;
-pub const FORMAT_BC2_UNORM_BLOCK: u32 = 135;
-pub const FORMAT_BC2_SRGB_BLOCK: u32 = 136;
-pub const FORMAT_BC3_UNORM_BLOCK: u32 = 137;
-pub const FORMAT_BC3_SRGB_BLOCK: u32 = 138;
-pub const FORMAT_BC4_UNORM_BLOCK: u32 = 139;
-pub const FORMAT_BC4_SNORM_BLOCK: u32 = 140;
-pub const FORMAT_BC5_UNORM_BLOCK: u32 = 141;
-pub const FORMAT_BC5_SNORM_BLOCK: u32 = 142;
-pub const FORMAT_BC6H_UFLOAT_BLOCK: u32 = 143;
-pub const FORMAT_BC6H_SFLOAT_BLOCK: u32 = 144;
-pub const FORMAT_BC7_UNORM_BLOCK: u32 = 145;
-pub const FORMAT_BC7_SRGB_BLOCK: u32 = 146;
-pub const FORMAT_ETC2_R8G8B8_UNORM_BLOCK: u32 = 147;
-pub const FORMAT_ETC2_R8G8B8_SRGB_BLOCK: u32 = 148;
-pub const FORMAT_ETC2_R8G8B8A1_UNORM_BLOCK: u32 = 149;
-pub const FORMAT_ETC2_R8G8B8A1_SRGB_BLOCK: u32 = 150;
-pub const FORMAT_ETC2_R8G8B8A8_UNORM_BLOCK: u32 = 151;
-pub const FORMAT_ETC2_R8G8B8A8_SRGB_BLOCK: u32 = 152;
-pub const FORMAT_EAC_R11_UNORM_BLOCK: u32 = 153;
-pub const FORMAT_EAC_R11_SNORM_BLOCK: u32 = 154;
-pub const FORMAT_EAC_R11G11_UNORM_BLOCK: u32 = 155;
-pub const FORMAT_EAC_R11G11_SNORM_BLOCK: u32 = 156;
-pub const FORMAT_ASTC_4x4_UNORM_BLOCK: u32 = 157;
-pub const FORMAT_ASTC_4x4_SRGB_BLOCK: u32 = 158;
-pub const FORMAT_ASTC_5x4_UNORM_BLOCK: u32 = 159;
-pub const FORMAT_ASTC_5x4_SRGB_BLOCK: u32 = 160;
-pub const FORMAT_ASTC_5x5_UNORM_BLOCK: u32 = 161;
-pub const FORMAT_ASTC_5x5_SRGB_BLOCK: u32 = 162;
-pub const FORMAT_ASTC_6x5_UNORM_BLOCK: u32 = 163;
-pub const FORMAT_ASTC_6x5_SRGB_BLOCK: u32 = 164;
-pub const FORMAT_ASTC_6x6_UNORM_BLOCK: u32 = 165;
-pub const FORMAT_ASTC_6x6_SRGB_BLOCK: u32 = 166;
-pub const FORMAT_ASTC_8x5_UNORM_BLOCK: u32 = 167;
-pub const FORMAT_ASTC_8x5_SRGB_BLOCK: u32 = 168;
-pub const FORMAT_ASTC_8x6_UNORM_BLOCK: u32 = 169;
-pub const FORMAT_ASTC_8x6_SRGB_BLOCK: u32 = 170;
-pub const FORMAT_ASTC_8x8_UNORM_BLOCK: u32 = 171;
-pub const FORMAT_ASTC_8x8_SRGB_BLOCK: u32 = 172;
-pub const FORMAT_ASTC_10x5_UNORM_BLOCK: u32 = 173;
-pub const FORMAT_ASTC_10x5_SRGB_BLOCK: u32 = 174;
-pub const FORMAT_ASTC_10x6_UNORM_BLOCK: u32 = 175;
-pub const FORMAT_ASTC_10x6_SRGB_BLOCK: u32 = 176;
-pub const FORMAT_ASTC_10x8_UNORM_BLOCK: u32 = 177;
-pub const FORMAT_ASTC_10x8_SRGB_BLOCK: u32 = 178;
-pub const FORMAT_ASTC_10x10_UNORM_BLOCK: u32 = 179;
-pub const FORMAT_ASTC_10x10_SRGB_BLOCK: u32 = 180;
-pub const FORMAT_ASTC_12x10_UNORM_BLOCK: u32 = 181;
-pub const FORMAT_ASTC_12x10_SRGB_BLOCK: u32 = 182;
-pub const FORMAT_ASTC_12x12_UNORM_BLOCK: u32 = 183;
-pub const FORMAT_ASTC_12x12_SRGB_BLOCK: u32 = 184;
-
-pub type ImageType = u32;
-pub const IMAGE_TYPE_1D: u32 = 0;
-pub const IMAGE_TYPE_2D: u32 = 1;
-pub const IMAGE_TYPE_3D: u32 = 2;
-
-pub type ImageTiling = u32;
-pub const IMAGE_TILING_OPTIMAL: u32 = 0;
-pub const IMAGE_TILING_LINEAR: u32 = 1;
-
-pub type PhysicalDeviceType = u32;
-pub const PHYSICAL_DEVICE_TYPE_OTHER: u32 = 0;
-pub const PHYSICAL_DEVICE_TYPE_INTEGRATED_GPU: u32 = 1;
-pub const PHYSICAL_DEVICE_TYPE_DISCRETE_GPU: u32 = 2;
-pub const PHYSICAL_DEVICE_TYPE_VIRTUAL_GPU: u32 = 3;
-pub const PHYSICAL_DEVICE_TYPE_CPU: u32 = 4;
-
-pub type QueryType = u32;
-pub const QUERY_TYPE_OCCLUSION: u32 = 0;
-pub const QUERY_TYPE_PIPELINE_STATISTICS: u32 = 1;
-pub const QUERY_TYPE_TIMESTAMP: u32 = 2;
-
-pub type SharingMode = u32;
-pub const SHARING_MODE_EXCLUSIVE: u32 = 0;
-pub const SHARING_MODE_CONCURRENT: u32 = 1;
-
-pub type ImageLayout = u32;
-pub const IMAGE_LAYOUT_UNDEFINED: u32 = 0;
-pub const IMAGE_LAYOUT_GENERAL: u32 = 1;
-pub const IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL: u32 = 2;
-pub const IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL: u32 = 3;
-pub const IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL: u32 = 4;
-pub const IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL: u32 = 5;
-pub const IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL: u32 = 6;
-pub const IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL: u32 = 7;
-pub const IMAGE_LAYOUT_PREINITIALIZED: u32 = 8;
-pub const IMAGE_LAYOUT_PRESENT_SRC_KHR: u32 = 1000001002;
-
-pub type ImageViewType = u32;
-pub const IMAGE_VIEW_TYPE_1D: u32 = 0;
-pub const IMAGE_VIEW_TYPE_2D: u32 = 1;
-pub const IMAGE_VIEW_TYPE_3D: u32 = 2;
-pub const IMAGE_VIEW_TYPE_CUBE: u32 = 3;
-pub const IMAGE_VIEW_TYPE_1D_ARRAY: u32 = 4;
-pub const IMAGE_VIEW_TYPE_2D_ARRAY: u32 = 5;
-pub const IMAGE_VIEW_TYPE_CUBE_ARRAY: u32 = 6;
-
-pub type ComponentSwizzle = u32;
-pub const COMPONENT_SWIZZLE_IDENTITY: u32 = 0;
-pub const COMPONENT_SWIZZLE_ZERO: u32 = 1;
-pub const COMPONENT_SWIZZLE_ONE: u32 = 2;
-pub const COMPONENT_SWIZZLE_R: u32 = 3;
-pub const COMPONENT_SWIZZLE_G: u32 = 4;
-pub const COMPONENT_SWIZZLE_B: u32 = 5;
-pub const COMPONENT_SWIZZLE_A: u32 = 6;
-
-pub type VertexInputRate = u32;
-pub const VERTEX_INPUT_RATE_VERTEX: u32 = 0;
-pub const VERTEX_INPUT_RATE_INSTANCE: u32 = 1;
-
-pub type PrimitiveTopology = u32;
-pub const PRIMITIVE_TOPOLOGY_POINT_LIST: u32 = 0;
-pub const PRIMITIVE_TOPOLOGY_LINE_LIST: u32 = 1;
-pub const PRIMITIVE_TOPOLOGY_LINE_STRIP: u32 = 2;
-pub const PRIMITIVE_TOPOLOGY_TRIANGLE_LIST: u32 = 3;
-pub const PRIMITIVE_TOPOLOGY_TRIANGLE_STRIP: u32 = 4;
-pub const PRIMITIVE_TOPOLOGY_TRIANGLE_FAN: u32 = 5;
-pub const PRIMITIVE_TOPOLOGY_LINE_LIST_WITH_ADJACENCY: u32 = 6;
-pub const PRIMITIVE_TOPOLOGY_LINE_STRIP_WITH_ADJACENCY: u32 = 7;
-pub const PRIMITIVE_TOPOLOGY_TRIANGLE_LIST_WITH_ADJACENCY: u32 = 8;
-pub const PRIMITIVE_TOPOLOGY_TRIANGLE_STRIP_WITH_ADJACENCY: u32 = 9;
-pub const PRIMITIVE_TOPOLOGY_PATCH_LIST: u32 = 10;
-
-pub type PolygonMode = u32;
-pub const POLYGON_MODE_FILL: u32 = 0;
-pub const POLYGON_MODE_LINE: u32 = 1;
-pub const POLYGON_MODE_POINT: u32 = 2;
-
-pub type FrontFace = u32;
-pub const FRONT_FACE_COUNTER_CLOCKWISE: u32 = 0;
-pub const FRONT_FACE_CLOCKWISE: u32 = 1;
-
-pub type CompareOp = u32;
-pub const COMPARE_OP_NEVER: u32 = 0;
-pub const COMPARE_OP_LESS: u32 = 1;
-pub const COMPARE_OP_EQUAL: u32 = 2;
-pub const COMPARE_OP_LESS_OR_EQUAL: u32 = 3;
-pub const COMPARE_OP_GREATER: u32 = 4;
-pub const COMPARE_OP_NOT_EQUAL: u32 = 5;
-pub const COMPARE_OP_GREATER_OR_EQUAL: u32 = 6;
-pub const COMPARE_OP_ALWAYS: u32 = 7;
-
-pub type StencilOp = u32;
-pub const STENCIL_OP_KEEP: u32 = 0;
-pub const STENCIL_OP_ZERO: u32 = 1;
-pub const STENCIL_OP_REPLACE: u32 = 2;
-pub const STENCIL_OP_INCREMENT_AND_CLAMP: u32 = 3;
-pub const STENCIL_OP_DECREMENT_AND_CLAMP: u32 = 4;
-pub const STENCIL_OP_INVERT: u32 = 5;
-pub const STENCIL_OP_INCREMENT_AND_WRAP: u32 = 6;
-pub const STENCIL_OP_DECREMENT_AND_WRAP: u32 = 7;
-
-pub type LogicOp = u32;
-pub const LOGIC_OP_CLEAR: u32 = 0;
-pub const LOGIC_OP_AND: u32 = 1;
-pub const LOGIC_OP_AND_REVERSE: u32 = 2;
-pub const LOGIC_OP_COPY: u32 = 3;
-pub const LOGIC_OP_AND_INVERTED: u32 = 4;
-pub const LOGIC_OP_NO_OP: u32 = 5;
-pub const LOGIC_OP_XOR: u32 = 6;
-pub const LOGIC_OP_OR: u32 = 7;
-pub const LOGIC_OP_NOR: u32 = 8;
-pub const LOGIC_OP_EQUIVALENT: u32 = 9;
-pub const LOGIC_OP_INVERT: u32 = 10;
-pub const LOGIC_OP_OR_REVERSE: u32 = 11;
-pub const LOGIC_OP_COPY_INVERTED: u32 = 12;
-pub const LOGIC_OP_OR_INVERTED: u32 = 13;
-pub const LOGIC_OP_NAND: u32 = 14;
-pub const LOGIC_OP_SET: u32 = 15;
-
-pub type BlendFactor = u32;
-pub const BLEND_FACTOR_ZERO: u32 = 0;
-pub const BLEND_FACTOR_ONE: u32 = 1;
-pub const BLEND_FACTOR_SRC_COLOR: u32 = 2;
-pub const BLEND_FACTOR_ONE_MINUS_SRC_COLOR: u32 = 3;
-pub const BLEND_FACTOR_DST_COLOR: u32 = 4;
-pub const BLEND_FACTOR_ONE_MINUS_DST_COLOR: u32 = 5;
-pub const BLEND_FACTOR_SRC_ALPHA: u32 = 6;
-pub const BLEND_FACTOR_ONE_MINUS_SRC_ALPHA: u32 = 7;
-pub const BLEND_FACTOR_DST_ALPHA: u32 = 8;
-pub const BLEND_FACTOR_ONE_MINUS_DST_ALPHA: u32 = 9;
-pub const BLEND_FACTOR_CONSTANT_COLOR: u32 = 10;
-pub const BLEND_FACTOR_ONE_MINUS_CONSTANT_COLOR: u32 = 11;
-pub const BLEND_FACTOR_CONSTANT_ALPHA: u32 = 12;
-pub const BLEND_FACTOR_ONE_MINUS_CONSTANT_ALPHA: u32 = 13;
-pub const BLEND_FACTOR_SRC_ALPHA_SATURATE: u32 = 14;
-pub const BLEND_FACTOR_SRC1_COLOR: u32 = 15;
-pub const BLEND_FACTOR_ONE_MINUS_SRC1_COLOR: u32 = 16;
-pub const BLEND_FACTOR_SRC1_ALPHA: u32 = 17;
-pub const BLEND_FACTOR_ONE_MINUS_SRC1_ALPHA: u32 = 18;
-
-pub type BlendOp = u32;
-pub const BLEND_OP_ADD: u32 = 0;
-pub const BLEND_OP_SUBTRACT: u32 = 1;
-pub const BLEND_OP_REVERSE_SUBTRACT: u32 = 2;
-pub const BLEND_OP_MIN: u32 = 3;
-pub const BLEND_OP_MAX: u32 = 4;
-
-pub type DynamicState = u32;
-pub const DYNAMIC_STATE_VIEWPORT: u32 = 0;
-pub const DYNAMIC_STATE_SCISSOR: u32 = 1;
-pub const DYNAMIC_STATE_LINE_WIDTH: u32 = 2;
-pub const DYNAMIC_STATE_DEPTH_BIAS: u32 = 3;
-pub const DYNAMIC_STATE_BLEND_CONSTANTS: u32 = 4;
-pub const DYNAMIC_STATE_DEPTH_BOUNDS: u32 = 5;
-pub const DYNAMIC_STATE_STENCIL_COMPARE_MASK: u32 = 6;
-pub const DYNAMIC_STATE_STENCIL_WRITE_MASK: u32 = 7;
-pub const DYNAMIC_STATE_STENCIL_REFERENCE: u32 = 8;
-
-pub type Filter = u32;
-pub const FILTER_NEAREST: u32 = 0;
-pub const FILTER_LINEAR: u32 = 1;
-
-pub type SamplerMipmapMode = u32;
-pub const SAMPLER_MIPMAP_MODE_NEAREST: u32 = 0;
-pub const SAMPLER_MIPMAP_MODE_LINEAR: u32 = 1;
-
-pub type SamplerAddressMode = u32;
-pub const SAMPLER_ADDRESS_MODE_REPEAT: u32 = 0;
-pub const SAMPLER_ADDRESS_MODE_MIRRORED_REPEAT: u32 = 1;
-pub const SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE: u32 = 2;
-pub const SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER: u32 = 3;
-pub const SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE: u32 = 4;
-
-pub type BorderColor = u32;
-pub const BORDER_COLOR_FLOAT_TRANSPARENT_BLACK: u32 = 0;
-pub const BORDER_COLOR_INT_TRANSPARENT_BLACK: u32 = 1;
-pub const BORDER_COLOR_FLOAT_OPAQUE_BLACK: u32 = 2;
-pub const BORDER_COLOR_INT_OPAQUE_BLACK: u32 = 3;
-pub const BORDER_COLOR_FLOAT_OPAQUE_WHITE: u32 = 4;
-pub const BORDER_COLOR_INT_OPAQUE_WHITE: u32 = 5;
-
-pub type DescriptorType = u32;
-pub const DESCRIPTOR_TYPE_SAMPLER: u32 = 0;
-pub const DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER: u32 = 1;
-pub const DESCRIPTOR_TYPE_SAMPLED_IMAGE: u32 = 2;
-pub const DESCRIPTOR_TYPE_STORAGE_IMAGE: u32 = 3;
-pub const DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER: u32 = 4;
-pub const DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER: u32 = 5;
-pub const DESCRIPTOR_TYPE_UNIFORM_BUFFER: u32 = 6;
-pub const DESCRIPTOR_TYPE_STORAGE_BUFFER: u32 = 7;
-pub const DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC: u32 = 8;
-pub const DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC: u32 = 9;
-pub const DESCRIPTOR_TYPE_INPUT_ATTACHMENT: u32 = 10;
-
-pub type AttachmentLoadOp = u32;
-pub const ATTACHMENT_LOAD_OP_LOAD: u32 = 0;
-pub const ATTACHMENT_LOAD_OP_CLEAR: u32 = 1;
-pub const ATTACHMENT_LOAD_OP_DONT_CARE: u32 = 2;
-
-pub type AttachmentStoreOp = u32;
-pub const ATTACHMENT_STORE_OP_STORE: u32 = 0;
-pub const ATTACHMENT_STORE_OP_DONT_CARE: u32 = 1;
-
-pub type PipelineBindPoint = u32;
-pub const PIPELINE_BIND_POINT_GRAPHICS: u32 = 0;
-pub const PIPELINE_BIND_POINT_COMPUTE: u32 = 1;
-
-pub type CommandBufferLevel = u32;
-pub const COMMAND_BUFFER_LEVEL_PRIMARY: u32 = 0;
-pub const COMMAND_BUFFER_LEVEL_SECONDARY: u32 = 1;
-
-pub type IndexType = u32;
-pub const INDEX_TYPE_UINT16: u32 = 0;
-pub const INDEX_TYPE_UINT32: u32 = 1;
-
-pub type SubpassContents = u32;
-pub const SUBPASS_CONTENTS_INLINE: u32 = 0;
-pub const SUBPASS_CONTENTS_SECONDARY_COMMAND_BUFFERS: u32 = 1;
-
-pub type InstanceCreateFlags = Flags;
-
-pub type FormatFeatureFlagBits = u32;
-pub const FORMAT_FEATURE_SAMPLED_IMAGE_BIT: u32 = 0x00000001;
-pub const FORMAT_FEATURE_STORAGE_IMAGE_BIT: u32 = 0x00000002;
-pub const FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT: u32 = 0x00000004;
-pub const FORMAT_FEATURE_UNIFORM_TEXEL_BUFFER_BIT: u32 = 0x00000008;
-pub const FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_BIT: u32 = 0x00000010;
-pub const FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_ATOMIC_BIT: u32 = 0x00000020;
-pub const FORMAT_FEATURE_VERTEX_BUFFER_BIT: u32 = 0x00000040;
-pub const FORMAT_FEATURE_COLOR_ATTACHMENT_BIT: u32 = 0x00000080;
-pub const FORMAT_FEATURE_COLOR_ATTACHMENT_BLEND_BIT: u32 = 0x00000100;
-pub const FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT: u32 = 0x00000200;
-pub const FORMAT_FEATURE_BLIT_SRC_BIT: u32 = 0x00000400;
-pub const FORMAT_FEATURE_BLIT_DST_BIT: u32 = 0x00000800;
-pub const FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT: u32 = 0x00001000;
-pub type FormatFeatureFlags = Flags;
-
-
-pub type ImageUsageFlagBits = u32;
-pub const IMAGE_USAGE_TRANSFER_SRC_BIT: u32 = 0x00000001;
-pub const IMAGE_USAGE_TRANSFER_DST_BIT: u32 = 0x00000002;
-pub const IMAGE_USAGE_SAMPLED_BIT: u32 = 0x00000004;
-pub const IMAGE_USAGE_STORAGE_BIT: u32 = 0x00000008;
-pub const IMAGE_USAGE_COLOR_ATTACHMENT_BIT: u32 = 0x00000010;
-pub const IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT: u32 = 0x00000020;
-pub const IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT: u32 = 0x00000040;
-pub const IMAGE_USAGE_INPUT_ATTACHMENT_BIT: u32 = 0x00000080;
-pub type ImageUsageFlags = Flags;
-
-
-pub type ImageCreateFlagBits = u32;
-pub const IMAGE_CREATE_SPARSE_BINDING_BIT: u32 = 0x00000001;
-pub const IMAGE_CREATE_SPARSE_RESIDENCY_BIT: u32 = 0x00000002;
-pub const IMAGE_CREATE_SPARSE_ALIASED_BIT: u32 = 0x00000004;
-pub const IMAGE_CREATE_MUTABLE_FORMAT_BIT: u32 = 0x00000008;
-pub const IMAGE_CREATE_CUBE_COMPATIBLE_BIT: u32 = 0x00000010;
-pub type ImageCreateFlags = Flags;
-
-
-pub type SampleCountFlagBits = u32;
-pub const SAMPLE_COUNT_1_BIT: u32 = 0x00000001;
-pub const SAMPLE_COUNT_2_BIT: u32 = 0x00000002;
-pub const SAMPLE_COUNT_4_BIT: u32 = 0x00000004;
-pub const SAMPLE_COUNT_8_BIT: u32 = 0x00000008;
-pub const SAMPLE_COUNT_16_BIT: u32 = 0x00000010;
-pub const SAMPLE_COUNT_32_BIT: u32 = 0x00000020;
-pub const SAMPLE_COUNT_64_BIT: u32 = 0x00000040;
-pub type SampleCountFlags = Flags;
-
-
-pub type QueueFlagBits = u32;
-pub const QUEUE_GRAPHICS_BIT: u32 = 0x00000001;
-pub const QUEUE_COMPUTE_BIT: u32 = 0x00000002;
-pub const QUEUE_TRANSFER_BIT: u32 = 0x00000004;
-pub const QUEUE_SPARSE_BINDING_BIT: u32 = 0x00000008;
-pub type QueueFlags = Flags;
-
-
-pub type MemoryPropertyFlagBits = u32;
-pub const MEMORY_PROPERTY_DEVICE_LOCAL_BIT: u32 = 0x00000001;
-pub const MEMORY_PROPERTY_HOST_VISIBLE_BIT: u32 = 0x00000002;
-pub const MEMORY_PROPERTY_HOST_COHERENT_BIT: u32 = 0x00000004;
-pub const MEMORY_PROPERTY_HOST_CACHED_BIT: u32 = 0x00000008;
-pub const MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT: u32 = 0x00000010;
-pub type MemoryPropertyFlags = Flags;
-
-
-pub type MemoryHeapFlagBits = u32;
-pub const MEMORY_HEAP_DEVICE_LOCAL_BIT: u32 = 0x00000001;
-pub type MemoryHeapFlags = Flags;
-pub type DeviceCreateFlags = Flags;
-pub type DeviceQueueCreateFlags = Flags;
-
-
-pub type PipelineStageFlagBits = u32;
-pub const PIPELINE_STAGE_TOP_OF_PIPE_BIT: u32 = 0x00000001;
-pub const PIPELINE_STAGE_DRAW_INDIRECT_BIT: u32 = 0x00000002;
-pub const PIPELINE_STAGE_VERTEX_INPUT_BIT: u32 = 0x00000004;
-pub const PIPELINE_STAGE_VERTEX_SHADER_BIT: u32 = 0x00000008;
-pub const PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT: u32 = 0x00000010;
-pub const PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT: u32 = 0x00000020;
-pub const PIPELINE_STAGE_GEOMETRY_SHADER_BIT: u32 = 0x00000040;
-pub const PIPELINE_STAGE_FRAGMENT_SHADER_BIT: u32 = 0x00000080;
-pub const PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT: u32 = 0x00000100;
-pub const PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT: u32 = 0x00000200;
-pub const PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT: u32 = 0x00000400;
-pub const PIPELINE_STAGE_COMPUTE_SHADER_BIT: u32 = 0x00000800;
-pub const PIPELINE_STAGE_TRANSFER_BIT: u32 = 0x00001000;
-pub const PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT: u32 = 0x00002000;
-pub const PIPELINE_STAGE_HOST_BIT: u32 = 0x00004000;
-pub const PIPELINE_STAGE_ALL_GRAPHICS_BIT: u32 = 0x00008000;
-pub const PIPELINE_STAGE_ALL_COMMANDS_BIT: u32 = 0x00010000;
-pub type PipelineStageFlags = Flags;
-pub type MemoryMapFlags = Flags;
-
-
-pub type ImageAspectFlagBits = u32;
-pub const IMAGE_ASPECT_COLOR_BIT: u32 = 0x00000001;
-pub const IMAGE_ASPECT_DEPTH_BIT: u32 = 0x00000002;
-pub const IMAGE_ASPECT_STENCIL_BIT: u32 = 0x00000004;
-pub const IMAGE_ASPECT_METADATA_BIT: u32 = 0x00000008;
-pub type ImageAspectFlags = Flags;
-
-
-pub type SparseImageFormatFlagBits = u32;
-pub const SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT: u32 = 0x00000001;
-pub const SPARSE_IMAGE_FORMAT_ALIGNED_MIP_SIZE_BIT: u32 = 0x00000002;
-pub const SPARSE_IMAGE_FORMAT_NONSTANDARD_BLOCK_SIZE_BIT: u32 = 0x00000004;
-pub type SparseImageFormatFlags = Flags;
-
-
-pub type SparseMemoryBindFlagBits = u32;
-pub const SPARSE_MEMORY_BIND_METADATA_BIT: u32 = 0x00000001;
-pub type SparseMemoryBindFlags = Flags;
-
-
-pub type FenceCreateFlagBits = u32;
-pub const FENCE_CREATE_SIGNALED_BIT: u32 = 0x00000001;
-pub type FenceCreateFlags = Flags;
-pub type SemaphoreCreateFlags = Flags;
-pub type EventCreateFlags = Flags;
-pub type QueryPoolCreateFlags = Flags;
-
-
-pub type QueryPipelineStatisticFlagBits = u32;
-pub const QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_VERTICES_BIT: u32 = 0x00000001;
-pub const QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_PRIMITIVES_BIT: u32 = 0x00000002;
-pub const QUERY_PIPELINE_STATISTIC_VERTEX_SHADER_INVOCATIONS_BIT: u32 = 0x00000004;
-pub const QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_INVOCATIONS_BIT: u32 = 0x00000008;
-pub const QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_PRIMITIVES_BIT: u32 = 0x00000010;
-pub const QUERY_PIPELINE_STATISTIC_CLIPPING_INVOCATIONS_BIT: u32 = 0x00000020;
-pub const QUERY_PIPELINE_STATISTIC_CLIPPING_PRIMITIVES_BIT: u32 = 0x00000040;
-pub const QUERY_PIPELINE_STATISTIC_FRAGMENT_SHADER_INVOCATIONS_BIT: u32 = 0x00000080;
-pub const QUERY_PIPELINE_STATISTIC_TESSELLATION_CONTROL_SHADER_PATCHES_BIT: u32 = 0x00000100;
-pub const QUERY_PIPELINE_STATISTIC_TESSELLATION_EVALUATION_SHADER_INVOCATIONS_BIT: u32 = 0x00000200;
-pub const QUERY_PIPELINE_STATISTIC_COMPUTE_SHADER_INVOCATIONS_BIT: u32 = 0x00000400;
-pub type QueryPipelineStatisticFlags = Flags;
-
-
-pub type QueryResultFlagBits = u32;
-pub const QUERY_RESULT_64_BIT: u32 = 0x00000001;
-pub const QUERY_RESULT_WAIT_BIT: u32 = 0x00000002;
-pub const QUERY_RESULT_WITH_AVAILABILITY_BIT: u32 = 0x00000004;
-pub const QUERY_RESULT_PARTIAL_BIT: u32 = 0x00000008;
-pub type QueryResultFlags = Flags;
-
-
-pub type BufferCreateFlagBits = u32;
-pub const BUFFER_CREATE_SPARSE_BINDING_BIT: u32 = 0x00000001;
-pub const BUFFER_CREATE_SPARSE_RESIDENCY_BIT: u32 = 0x00000002;
-pub const BUFFER_CREATE_SPARSE_ALIASED_BIT: u32 = 0x00000004;
-pub type BufferCreateFlags = Flags;
-
-
-pub type BufferUsageFlagBits = u32;
-pub const BUFFER_USAGE_TRANSFER_SRC_BIT: u32 = 0x00000001;
-pub const BUFFER_USAGE_TRANSFER_DST_BIT: u32 = 0x00000002;
-pub const BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT: u32 = 0x00000004;
-pub const BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT: u32 = 0x00000008;
-pub const BUFFER_USAGE_UNIFORM_BUFFER_BIT: u32 = 0x00000010;
-pub const BUFFER_USAGE_STORAGE_BUFFER_BIT: u32 = 0x00000020;
-pub const BUFFER_USAGE_INDEX_BUFFER_BIT: u32 = 0x00000040;
-pub const BUFFER_USAGE_VERTEX_BUFFER_BIT: u32 = 0x00000080;
-pub const BUFFER_USAGE_INDIRECT_BUFFER_BIT: u32 = 0x00000100;
-pub type BufferUsageFlags = Flags;
-pub type BufferViewCreateFlags = Flags;
-pub type ImageViewCreateFlags = Flags;
-pub type ShaderModuleCreateFlags = Flags;
-pub type PipelineCacheCreateFlags = Flags;
-
-
-pub type PipelineCreateFlagBits = u32;
-pub const PIPELINE_CREATE_DISABLE_OPTIMIZATION_BIT: u32 = 0x00000001;
-pub const PIPELINE_CREATE_ALLOW_DERIVATIVES_BIT: u32 = 0x00000002;
-pub const PIPELINE_CREATE_DERIVATIVE_BIT: u32 = 0x00000004;
-pub type PipelineCreateFlags = Flags;
-pub type PipelineShaderStageCreateFlags = Flags;
-
-
-pub type ShaderStageFlagBits = u32;
-pub const SHADER_STAGE_VERTEX_BIT: u32 = 0x00000001;
-pub const SHADER_STAGE_TESSELLATION_CONTROL_BIT: u32 = 0x00000002;
-pub const SHADER_STAGE_TESSELLATION_EVALUATION_BIT: u32 = 0x00000004;
-pub const SHADER_STAGE_GEOMETRY_BIT: u32 = 0x00000008;
-pub const SHADER_STAGE_FRAGMENT_BIT: u32 = 0x00000010;
-pub const SHADER_STAGE_COMPUTE_BIT: u32 = 0x00000020;
-pub const SHADER_STAGE_ALL_GRAPHICS: u32 = 0x1F;
-pub const SHADER_STAGE_ALL: u32 = 0x7FFFFFFF;
-pub type PipelineVertexInputStateCreateFlags = Flags;
-pub type PipelineInputAssemblyStateCreateFlags = Flags;
-pub type PipelineTessellationStateCreateFlags = Flags;
-pub type PipelineViewportStateCreateFlags = Flags;
-pub type PipelineRasterizationStateCreateFlags = Flags;
-
-
-pub type CullModeFlagBits = u32;
-pub const CULL_MODE_NONE: u32 = 0;
-pub const CULL_MODE_FRONT_BIT: u32 = 0x00000001;
-pub const CULL_MODE_BACK_BIT: u32 = 0x00000002;
-pub const CULL_MODE_FRONT_AND_BACK: u32 = 0x3;
-pub type CullModeFlags = Flags;
-pub type PipelineMultisampleStateCreateFlags = Flags;
-pub type PipelineDepthStencilStateCreateFlags = Flags;
-pub type PipelineColorBlendStateCreateFlags = Flags;
-
-
-pub type ColorComponentFlagBits = u32;
-pub const COLOR_COMPONENT_R_BIT: u32 = 0x00000001;
-pub const COLOR_COMPONENT_G_BIT: u32 = 0x00000002;
-pub const COLOR_COMPONENT_B_BIT: u32 = 0x00000004;
-pub const COLOR_COMPONENT_A_BIT: u32 = 0x00000008;
-pub type ColorComponentFlags = Flags;
-pub type PipelineDynamicStateCreateFlags = Flags;
-pub type PipelineLayoutCreateFlags = Flags;
-pub type ShaderStageFlags = Flags;
-pub type SamplerCreateFlags = Flags;
-pub type DescriptorSetLayoutCreateFlags = Flags;
-
-
-pub type DescriptorPoolCreateFlagBits = u32;
-pub const DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT: u32 = 0x00000001;
-pub type DescriptorPoolCreateFlags = Flags;
-pub type DescriptorPoolResetFlags = Flags;
-pub type FramebufferCreateFlags = Flags;
-pub type RenderPassCreateFlags = Flags;
-
-
-pub type AttachmentDescriptionFlagBits = u32;
-pub const ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT: u32 = 0x00000001;
-pub type AttachmentDescriptionFlags = Flags;
-pub type SubpassDescriptionFlags = Flags;
-
-
-pub type AccessFlagBits = u32;
-pub const ACCESS_INDIRECT_COMMAND_READ_BIT: u32 = 0x00000001;
-pub const ACCESS_INDEX_READ_BIT: u32 = 0x00000002;
-pub const ACCESS_VERTEX_ATTRIBUTE_READ_BIT: u32 = 0x00000004;
-pub const ACCESS_UNIFORM_READ_BIT: u32 = 0x00000008;
-pub const ACCESS_INPUT_ATTACHMENT_READ_BIT: u32 = 0x00000010;
-pub const ACCESS_SHADER_READ_BIT: u32 = 0x00000020;
-pub const ACCESS_SHADER_WRITE_BIT: u32 = 0x00000040;
-pub const ACCESS_COLOR_ATTACHMENT_READ_BIT: u32 = 0x00000080;
-pub const ACCESS_COLOR_ATTACHMENT_WRITE_BIT: u32 = 0x00000100;
-pub const ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT: u32 = 0x00000200;
-pub const ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT: u32 = 0x00000400;
-pub const ACCESS_TRANSFER_READ_BIT: u32 = 0x00000800;
-pub const ACCESS_TRANSFER_WRITE_BIT: u32 = 0x00001000;
-pub const ACCESS_HOST_READ_BIT: u32 = 0x00002000;
-pub const ACCESS_HOST_WRITE_BIT: u32 = 0x00004000;
-pub const ACCESS_MEMORY_READ_BIT: u32 = 0x00008000;
-pub const ACCESS_MEMORY_WRITE_BIT: u32 = 0x00010000;
-pub type AccessFlags = Flags;
-
-
-pub type DependencyFlagBits = u32;
-pub const DEPENDENCY_BY_REGION_BIT: u32 = 0x00000001;
-pub type DependencyFlags = Flags;
-
-
-pub type CommandPoolCreateFlagBits = u32;
-pub const COMMAND_POOL_CREATE_TRANSIENT_BIT: u32 = 0x00000001;
-pub const COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT: u32 = 0x00000002;
-pub type CommandPoolCreateFlags = Flags;
-
-
-pub type CommandPoolResetFlagBits = u32;
-pub const COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT: u32 = 0x00000001;
-pub type CommandPoolResetFlags = Flags;
-
-
-pub type CommandBufferUsageFlagBits = u32;
-pub const COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT: u32 = 0x00000001;
-pub const COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT: u32 = 0x00000002;
-pub const COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT: u32 = 0x00000004;
-pub type CommandBufferUsageFlags = Flags;
-
-
-pub type QueryControlFlagBits = u32;
-pub const QUERY_CONTROL_PRECISE_BIT: u32 = 0x00000001;
-pub type QueryControlFlags = Flags;
-
-
-pub type CommandBufferResetFlagBits = u32;
-pub const COMMAND_BUFFER_RESET_RELEASE_RESOURCES_BIT: u32 = 0x00000001;
-pub type CommandBufferResetFlags = Flags;
-
-
-pub type StencilFaceFlagBits = u32;
-pub const STENCIL_FACE_FRONT_BIT: u32 = 0x00000001;
-pub const STENCIL_FACE_BACK_BIT: u32 = 0x00000002;
-pub const STENCIL_FRONT_AND_BACK: u32 = 0x3;
-pub type StencilFaceFlags = Flags;
-
-
-pub type DisplayPlaneAlphaFlagBitsKHR = u32;
-pub const DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR: u32 = 0x00000001;
-pub const DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR: u32 = 0x00000002;
-pub const DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR: u32 = 0x00000004;
-pub const DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR: u32 = 0x00000008;
-pub type DisplayModeCreateFlagsKHR = Flags;
-pub type DisplayPlaneAlphaFlagsKHR = Flags;
-pub type DisplaySurfaceCreateFlagsKHR = Flags;
-
-pub type ColorSpaceKHR = u32;
-pub const COLORSPACE_SRGB_NONLINEAR_KHR: u32 = 0;
-
-pub type PresentModeKHR = u32;
-pub const PRESENT_MODE_IMMEDIATE_KHR: u32 = 0;
-pub const PRESENT_MODE_MAILBOX_KHR: u32 = 1;
-pub const PRESENT_MODE_FIFO_KHR: u32 = 2;
-pub const PRESENT_MODE_FIFO_RELAXED_KHR: u32 = 3;
-
-pub type SurfaceTransformFlagBitsKHR = u32;
-pub const SURFACE_TRANSFORM_IDENTITY_BIT_KHR: u32 = 0x00000001;
-pub const SURFACE_TRANSFORM_ROTATE_90_BIT_KHR: u32 = 0x00000002;
-pub const SURFACE_TRANSFORM_ROTATE_180_BIT_KHR: u32 = 0x00000004;
-pub const SURFACE_TRANSFORM_ROTATE_270_BIT_KHR: u32 = 0x00000008;
-pub const SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR: u32 = 0x00000010;
-pub const SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR: u32 = 0x00000020;
-pub const SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR: u32 = 0x00000040;
-pub const SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR: u32 = 0x00000080;
-pub const SURFACE_TRANSFORM_INHERIT_BIT_KHR: u32 = 0x00000100;
-pub type SurfaceTransformFlagsKHR = Flags;
-
-pub type CompositeAlphaFlagBitsKHR = u32;
-pub const COMPOSITE_ALPHA_OPAQUE_BIT_KHR: u32 = 0x00000001;
-pub const COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR: u32 = 0x00000002;
-pub const COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR: u32 = 0x00000004;
-pub const COMPOSITE_ALPHA_INHERIT_BIT_KHR: u32 = 0x00000008;
-pub type CompositeAlphaFlagsKHR = Flags;
-
-pub type DebugReportObjectTypeEXT = u32;
-pub const DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT: u32 = 0;
-pub const DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT: u32 = 1;
-pub const DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT: u32 = 2;
-pub const DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT: u32 = 3;
-pub const DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT: u32 = 4;
-pub const DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT: u32 = 5;
-pub const DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT: u32 = 6;
-pub const DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT: u32 = 7;
-pub const DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT: u32 = 8;
-pub const DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT: u32 = 9;
-pub const DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT: u32 = 10;
-pub const DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT: u32 = 11;
-pub const DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT: u32 = 12;
-pub const DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT: u32 = 13;
-pub const DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT: u32 = 14;
-pub const DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT: u32 = 15;
-pub const DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT: u32 = 16;
-pub const DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT: u32 = 17;
-pub const DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT: u32 = 18;
-pub const DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT: u32 = 19;
-pub const DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT: u32 = 20;
-pub const DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT: u32 = 21;
-pub const DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT: u32 = 22;
-pub const DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT: u32 = 23;
-pub const DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT: u32 = 24;
-pub const DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT: u32 = 25;
-pub const DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT: u32 = 26;
-pub const DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT: u32 = 27;
-pub const DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_EXT: u32 = 28;
-
-pub type DebugReportErrorEXT = u32;
-pub const DEBUG_REPORT_ERROR_NONE_EXT: u32 = 0;
-pub const DEBUG_REPORT_ERROR_CALLBACK_REF_EXT: u32 = 1;
-
-pub type DebugReportFlagBitsEXT = u32;
-pub const DEBUG_REPORT_INFORMATION_BIT_EXT: u32 = 0x00000001;
-pub const DEBUG_REPORT_WARNING_BIT_EXT: u32 = 0x00000002;
-pub const DEBUG_REPORT_PERFORMANCE_WARNING_BIT_EXT: u32 = 0x00000004;
-pub const DEBUG_REPORT_ERROR_BIT_EXT: u32 = 0x00000008;
-pub const DEBUG_REPORT_DEBUG_BIT_EXT: u32 = 0x00000010;
-pub type DebugReportFlagsEXT = Flags;
